drafted new ModelSection object, general encapsulation of sections in core.py
