+++ .gitignore
@@ -14,3 +14,6 @@ __pycache__/
 .cache/
 .pytest_cache/
 _build/
+
+# IDE Stuff
+*.idea/
