+++ swmmio/defs/section_headers.json
@@ -1,0 +1,5 @@
+{
+  "junctions": ["Name", "Elevation", "MaxDepth", "InitDepth", "SurDepth", "Aponded"],
+  "outfalls": ["Name", "Elevation", "Type", "Stage Data", "Gated", "Route To"],
+  "conduits": ["Name", "InletNode", "OutletNode", "Length", "ManningN", "InletOffset", "OutletOffset", "InitFlow", "MaxFlow"]
+}
